third_party/blink/web_tests/platform/generic/fast/dom/HTMLDocument/object-by-name-or-id-expected.txt

This tests when object elements are accessible by name or id directly as properties of the document object. A PASS result means that behavior matches IE and Firefox.

Results:

By name:
no children: PASS
param: PASS
param and whitespace: PASS
param and empty comment: PASS
param and non-empty comment: PASS
param and text: PASS
param and img: PASS

By id:
no children: PASS
param: PASS
param and whitespace: PASS
param and empty comment: PASS
param and non-empty comment: PASS
param and text: PASS
param and img: PASS
